requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
dj-database-url==0.5.0
Django==2.2
django-cors-headers==2.5.2
djangorestframework==3.9.2
idna==3.3
python-decouple==3.1
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
